uses local code for docker, enables env vars for config
This commit removes the git command embedded in the Dockerfile
in favor of using Docker's COPY, eliminating the need to download the
repo twice. To accommodate this change I had to move the location of
the bundle.js creation to the root of the repository and commit a
relative symlink to the repository.